examples/vertex-ai-foundations/blueprints/third-party-solutions/openshift/requirements.txt
click
pyhcl
ruamel.yaml